Add Desktop Compose UI counter frontend
