+++ gradle/libs.versions.toml
@@ -4,10 +4,8 @@ kotlinx-coroutines = "1.6.4"
 kotlinx-serialization = "1.4.1"
 androidxComposeCompiler = "1.3.2"
 androidx-activity = "1.3.1"
-androidx-compose = "1.2.1"
+jbCompose = "1.2.2"
 jbComposeCompiler = "1.3.2.2"
-jbComposeRuntime = "1.2.2"
-jbComposeFoundation = "1.2.2"
 lint = "30.4.0-alpha09"
 zipline = "0.9.12"
 coil = "2.2.2"
@@ -37,10 +35,6 @@ androidx-activity = { module = "androidx.activity:activity", version.ref = "andr
 androidx-activity-compose = { module = "androidx.activity:activity-compose", version.ref = "androidx.activity" }
 androidx-appCompat = { module = "androidx.appcompat:appcompat", version = "1.3.1" }
 androidx-core = { module = "androidx.core:core-ktx", version = "1.8.0" }
-androidx-compose-material = { module = "androidx.compose.material:material", version.ref = "androidx-compose" }
-androidx-compose-ui = { module = "androidx.compose.ui:ui", version.ref = "androidx-compose" }
-androidx-compose-ui-tooling = { module = "androidx.compose.ui:ui-tooling", version.ref = "androidx-compose" }
-androidx-compose-ui-tooling-preview = { module = "androidx.compose.ui:ui-tooling-preview", version.ref = "androidx-compose" }
 androidx-recyclerview = { module = "androidx.recyclerview:recyclerview", version = "1.2.1" }
 androidx-test-core = "androidx.test:core-ktx:1.5.0"
 androidx-test-runner = "androidx.test:runner:1.5.1"
@@ -48,8 +42,13 @@ androidx-test-uiautomator = "androidx.test.uiautomator:uiautomator:2.2.0"
 
 google-material = { module = "com.google.android.material:material", version = "1.4.0" }
 
-jetbrains-compose-runtime = { module = "org.jetbrains.compose.runtime:runtime", version.ref = "jbComposeRuntime" }
-jetbrains-compose-foundation = { module = "org.jetbrains.compose.foundation:foundation", version.ref = "jbComposeFoundation" }
+jetbrains-compose-gradlePlugin = { module = "org.jetbrains.compose:compose-gradle-plugin", version.ref = "jbCompose" }
+jetbrains-compose-foundation = { module = "org.jetbrains.compose.foundation:foundation", version.ref = "jbCompose" }
+jetbrains-compose-material = { module = "org.jetbrains.compose.material:material", version.ref = "jbCompose" }
+jetbrains-compose-runtime = { module = "org.jetbrains.compose.runtime:runtime", version.ref = "jbCompose" }
+jetbrains-compose-ui = { module = "org.jetbrains.compose.ui:ui", version.ref = "jbCompose" }
+jetbrains-compose-ui-tooling = { module = "org.jetbrains.compose.ui:ui-tooling", version.ref = "jbCompose" }
+jetbrains-compose-ui-tooling-preview = { module = "org.jetbrains.compose.ui:ui-tooling-preview", version.ref = "jbCompose" }
 
 androidGradlePlugin = { module = "com.android.tools.build:gradle", version = "7.3.1" }
 kotlinPoet = { module = "com.squareup:kotlinpoet", version = "1.12.0" }
