+++ samples/counter/android-composeui/build.gradle
@@ -3,19 +3,18 @@ apply plugin: 'org.jetbrains.kotlin.android'
 apply plugin: 'app.cash.redwood'
 
 dependencies {
-  implementation projects.samples.counter.sunspot.widget
   implementation projects.samples.counter.presenter
+  implementation projects.samples.counter.sharedComposeui
+  implementation projects.redwoodWidgetCompose
   implementation libs.kotlinx.coroutines.android
-
   implementation libs.google.material
   implementation libs.androidx.activity.compose
   implementation libs.androidx.appCompat
   implementation libs.androidx.core
-  implementation libs.androidx.compose.material
-  implementation libs.androidx.compose.ui
-  implementation libs.androidx.compose.ui.tooling.preview
-  implementation projects.redwoodWidgetCompose
-  debugImplementation libs.androidx.compose.ui.tooling
+  implementation libs.jetbrains.compose.material
+  implementation libs.jetbrains.compose.ui
+  implementation libs.jetbrains.compose.ui.tooling.preview
+  debugImplementation libs.jetbrains.compose.ui.tooling
 
   androidTestImplementation libs.androidx.test.runner
   androidTestImplementation projects.samples.counter.androidTests
