--- samples/counter/android-composeui/src/main/java/example/android/sunspot/AndroidSunspotBox.kt
@@ -1,41 +1,0 @@
-/*
- * Copyright (C) 2022 Square, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package example.android.sunspot
-
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.padding
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
-import app.cash.redwood.LayoutModifier
-import app.cash.redwood.widget.compose.ComposeWidgetChildren
-import example.sunspot.widget.SunspotBox
-
-class AndroidSunspotBox : SunspotBox<@Composable () -> Unit> {
-  override var layoutModifiers: LayoutModifier = LayoutModifier
-
-  override val children = ComposeWidgetChildren()
-
-  override val value = @Composable {
-    Column(
-      horizontalAlignment = Alignment.CenterHorizontally,
-      modifier = Modifier.padding(16.dp)
-    ) {
-      children.render()
-    }
-  }
-}
