+++ samples/emoji-search/android-composeui/build.gradle
@@ -22,12 +22,12 @@ dependencies {
   implementation libs.androidx.activity.compose
   implementation libs.androidx.appCompat
   implementation libs.androidx.core
-  implementation libs.androidx.compose.material
-  implementation libs.androidx.compose.ui
-  implementation libs.androidx.compose.ui.tooling.preview
+  implementation libs.jetbrains.compose.material
+  implementation libs.jetbrains.compose.ui
+  implementation libs.jetbrains.compose.ui.tooling.preview
+  debugImplementation libs.jetbrains.compose.ui.tooling
   implementation projects.redwoodLayoutComposeui
   implementation projects.redwoodTreehouseLazylayoutComposeui
   implementation projects.redwoodTreehouseComposeui
   implementation projects.redwoodWidgetCompose
-  debugImplementation libs.androidx.compose.ui.tooling
 }
