+++ settings.gradle
@@ -41,12 +41,14 @@ include ':test-schema:widget'
 include ':test-schema:widget:protocol'
 
 if (!hasProperty('redwoodNoSamples')) {
+  include ':samples:counter:android-composeui'
   include ':samples:counter:android-views'
   include ':samples:counter:android-tests'
-  include ':samples:counter:android-composeui'
   include ':samples:counter:browser'
+  include ':samples:counter:desktop-composeui'
   include ':samples:counter:ios-shared'
   include ':samples:counter:presenter'
+  include ':samples:counter:shared-composeui'
   include ':samples:counter:sunspot'
   include ':samples:counter:sunspot:compose'
   include ':samples:counter:sunspot:widget'
